Add functionality to operator buttons
The calculator now has complete functionality. Each operator button
has been tested to work correctly and can be used to perform
operations in sequence.

e.g.: 12 + 5 - 1 * 2 = 32

The display updates as operations are entered, so in the above example,
it would display, in order, 12, 5, 17, 1, 16, 2, and 32. Operators
are not displayed in the calculator output, though may be added in the
future.

The equals button computes a result, but only if enough information has
been entered. The result of pressing equals can then be used in the next
operation.

The clear button resets all elements to their starting values.

Each result is trimmed to five decimal places before being displayed. If
a value is too long to fit on the screen or if a divide by 0 is
attempted, it displays "ERROR".